Bug with new Windows function removed. Bug with Choosing Colour removed.
Bug with new Windows function removed:
There was a bug if the text was longer, it overflows the whole window. Added new height variables and also a scrollbar.

Bug with Choosing Colour removed:
If Player choosed other Colour, the Coin Animation was incorrect and also there was a problem with storing  the Colour Checkbox value correct and show the correect value after Page refresh. Solved that with local Storage & Game Object storing like in the Audio Section.